Completed NPCController
