+++ Assets/Scripts/NPC.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 2f1beb424e51de248a6e4137d7c50e88
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
