+++ Assets/Scripts/NPC/NPCObjects.cs.meta
