--- Assets/Scripts/NPCObjects.cs
@@ -1,33 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-
-//READ THIS: MOST, IF NOT ALL OF THE PUBLIC VARIABLES ARE BEING ASSIGNED TO THE GREEN CIRCLE NPCPOINT!
-
-
-public class NPCObjects : MonoBehaviour
-{
-    public bool occupied = false;
-
-    // Start is called before the first frame update
-    void Start()
-    {
-
-    }
-
-    // Update is called once per frame
-    void Update()
-    {
-        
-    }
-}
-
-
-/*
- * TODO:
- *  delete unused functions
- *  write update variable function
- *  seat manager + npc manager 
- *  put all seat objects under parent object (so that i can loop through) 
-*/
